requirements.txt
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.11.0
blinker==1.6.2
bs4==0.0.1
certifi==2023.5.7
cffi==1.15.1
cfgv==3.4.0
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
distlib==0.3.7
et-xmlfile==1.1.0
exceptiongroup==1.1.1
filelock==3.13.1
Flask==2.3.2
Flask-WTF==1.1.1
fonttools==4.49.0
h11==0.14.0
identify==2.5.32
idna==3.4
imageio==2.31.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
lazy_loader==0.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
networkx==3.1
nodeenv==1.8.0
numpy==1.24.3
opencv-contrib-python==4.9.0.80
opencv-python==4.7.0.72
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.0.1
pathspec==0.11.2
Pillow==9.5.0
platformdirs==4.0.0
pre-commit==3.5.0
pycparser==2.21
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyWavelets==1.4.1
# pywin32==306
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.21.0
scipy==1.10.1
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tifffile==2023.4.12
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
types-python-dateutil==2.8.19.14
tzdata==2023.3
urllib3==2.0.2
virtualenv==20.24.7
wcwidth==0.2.8
webdriver-manager==3.8.6
webencodings==0.5.1
Werkzeug==2.3.6
wsproto==1.2.0
WTForms==3.0.1
